requirements.txt
aiosqlite==0.17.0
anyio==4.1.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
asyncpg==0.29.0
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
comm==0.2.0
coreapi==2.3.3
coreschema==0.0.4
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.21
django-environ==0.12.0
django-filter==25.1
django-guardian==3.0.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
executing==2.0.1
fastapi==0.115.8
fastjsonschema==2.19.0
Flask==3.0.1
fqdn==1.5.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.6
importlib_metadata==7.1.0
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.27.1
ipython==8.18.1
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.1
iso8601==1.1.0
isoduration==20.11.0
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.11.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
nats-py==2.5.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==2.0.0
openapi-codec==1.3.2
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asyncpg==0.46b0
opentelemetry-instrumentation-tortoiseorm==0.46b0
opentelemetry-propagator-jaeger==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
overrides==7.4.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.6
pexpect==4.9.0
platformdirs==4.0.0
pluggy==1.5.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.25.3
psutil==5.9.6
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycurl==7.45.2
pydantic==1.10.0
Pygments==2.17.2
PyJWT==2.8.0
pypika-tortoise==0.1.6
pytest==8.1.1
pytest-asyncio==0.23.7
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.31.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.1
scipy==1.14.0
Send2Trash==1.8.2
setuptools==70.1.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.45.3
statsmodels==0.14.0
terminado==0.18.0
tinycss2==1.2.1
tornado==6.3.3
tortoise-orm==0.19.3
traitlets==5.14.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.34.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.1.9
xmltodict==0.13.0
zipp==3.19.2
